requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
einops==0.7.0
encodec==0.1.1
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
g2pkk==0.1.2
grpcio==1.62.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.22.2
idna==3.7
inflect==7.2.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.0
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.22.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==24.0
pandas==1.5.3
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
preshed==3.0.9
protobuf==5.26.1
psutil==5.9.8
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
srt==3.5.3
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
tqdm==4.66.2
trainer==0.0.36
transformers==4.40.0
triton==2.2.0
TTS==0.22.0
typeguard==4.2.1
typer==0.9.4
typing_extensions==4.11.0
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
vosk==0.3.45
wasabi==1.1.2
Wave==0.0.2
weasel==0.3.4
websockets==12.0
Werkzeug==3.0.2
yarl==1.9.4
